Test Spec Structure Updated and Added New Specs (8 out of 10 are working)
