+++ package.json
@@ -20,7 +20,7 @@
     "start": "node server",
     "start-dev": "NODE_ENV='development' npm run build-client-watch & NODE_ENV='development' npm run start-server",
     "start-server": "nodemon server -e html,js,scss --ignore public --ignore client",
-    "test": "NODE_ENV='test' mocha \"./server/**/*.spec.js\" \"./client/**/*.spec.js\" \"./script/**/*.spec.js\" --require @babel/polyfill --require @babel/register",
+    "test": "NODE_ENV='test' mocha \"./server/**/*.spec.js\" \"./client/**/*.spec.js\" \"./script/**/*.spec.js\" --recursive --watch --require @babel/polyfill --require @babel/register",
     "test2": "NODE_ENV='test' mocha  ./test \"./test/tier1/*.spec.js\" --recursive --watch --require @babel/polyfill --require @babel/register"
   },
   "husky": {
